[mux] the world's most over-engineered mux is done
Sweet! That was a tough night of setting up all the tools, but for now,
we're looking pretty good. Installed gtk-wave, iverilog, and a whole
bunch of random other packages.

To test the mux, go to `tests/mux` and run `make`. Tests pass! Also
verified in GTK wave since I'm not sure what I'm doing for now...